Added (broken) create_process() syscall

+++ kernel/palloc.cc
@@ -45,7 +45,7 @@ void* palloc(size_t size)
         arg.writeable = 1;
         arg.execution_disabled = 1;
 
-        for (int i = 0; i < size_p; i += 4096) {
+        for (uint64_t i = 0; i < size_p; i += 4096) {
             get_page(p_head + i, arg);
         }
